cmocka tpl, utilities with unit tests, start sonar
Start work on sonar program. Configure USART for the MaxBotix
LV-MaxSonar-EZ MB1010. Need to work on interupts and receiving the sonar
range value.

Add cmocka as a third party library. Break out some utility functions
for common bit operations and wrappers for turning the led on and off
and changing the color. Unit test utility functions.