Change approach to turning off led during sleep
Setting the output pin to zero logic level doesn't necessarily set the voltage
to ground. The remaining current flowing through the pin is enough to light the
led. As a result the original method of turning off the led by setting the
anode pin to zero doesn't work unless a transistor is attached to the pin to
sink the line to ground. For now, add a new led_off method that sets the anode
to zero and all the cathodes to one. Then the resulting current can't flow
through the led and it will turn off.

Testing: unit tests pass.