Add timer to to monitor changes in range
Setup the 16 bit timer / counter 1 and an additional software counter as
a way to monitor the changes in range. Every time the delta_range
exceeds a cutoff, we reset the timers. When the hardware counter
overflows, it triggers an interrupt that increments the software
counter. If the range is in a steady state, we go to sleep.

Testing: unit tests pass, manual sonar testing ok.