Bugfix: allow sonar a complete ranging cycle upon wakeup
The event loop is very short, and it's likely that we are
trying to turn the sonar on/off in the middle of a ranging cycle...?
Setting a delay to ensure one complete cycle ensures we get a new data
point and seems to prevent a bug where the sonar gets stuck at the
minimum range.

Testing: manually tested sonar, led, sleep.